chore(api): Replace Model#getId() with handling foreign key constraint error
Also as part of this PR, i have updated the POST /slots endpoint to
support bulk data and added some tests for it.
The `capacity` was also added as required request param in that endpoint
as it should never be stored as null and there is no default value at
the momemnt. We could pass the default value from shop settings in
the future and loosen that restriction'.